Add support for an idle timeout when mounting a squashfile.
If specified via the timeout option (-o timeout=N), then squashfuse_ll
will track the time of all accesses as well as the count of open files
and directories.  If the open count is zero and more seconds than the
specified timeout have passed since the last activity on the
filesystem, then we exit.

The approach is based on fuse's own signal handling (from
lib/fuse_signals.c).