Support libfuse version 3.
Version 3 of libfuse broke backwards API compatibility (for good
reasons). See https://github.com/libfuse/libfuse/releases/tag/fuse-3.0.0

Use autoconf to determine if libfuse3 is available (and is at least
version 3.2, which is the least version tested). If so, wield the
new APIs.

The main motivation is the PARALLEL_DIROPS feature as well as general
future-proofing.