Add locking to cache API in anticipation of multithreaded decompression.
The existing cache API has effectively no locking and relies on the single-
threaded nature of the code to prevent contention on cache entries.
As a first step to multithreading the driver, refactor the cache API to allow
alternative implementations. In this changeset the API is changed but the
internal cache implementation is not make thread-safe, so it still suitable
only for single-threaded usage.

Specific changes:

*   sqfs_cache type is made opaque.
*   previously, newly allocated cache entries were assumed valid. This meant that
    on any failure path following a cache entry allocation, one had to be careful
    to call sqfs_cache_invalidate(). The assumption is now reversed, cache
    entries are invalid until explicitly marked valid with
    sqfs_cache_entry_valid(). This simplifies error handling.
*   block cache code was intermixed with generic cache code, relocate to fs.c
*   cache eviction led to object destruction in block cache. The only thing
    preventing cache eviction while block object in use is single-threaded
    nature of code. Instead use a refcounting mechanism on the block entries
    so that we can independently manage block lifetime.