Add a notification mechanism for signaling when the mountpoint is ready
Add the notify_pipe option which allows the user to specify the path to
a named pipe. When the mountpoint is ready, a message will be written to
this pipe: 's' for success and 'f' for failure. To avoid blocking the
fuse process until the pipe is read, a new process is spawned which
performs a blocking operation on the pipe. In case of mount failure, no
additional process is created and the main process blocks until the pipe
is read.

An example of operation is provided below:
```
set -x
FIFO=$(mktemp -u)
mkfifo "$FIFO"
./squashfuse_ll -o notify_pipe="$FIFO" -f /path/to/squashfs/archive /tmp/squash&
STATUS=$(head -c1 "$FIFO")
if [ "$STATUS" = "s" ]; then
	echo "Mountpoint contains:"
	ls /tmp/squash
else
	echo "Mounting squashfuse on /tmp/squash failed"
fi
```

Fixes #49